6-size.c

+++ 0x00-hello_world/6-size.c
@@ -14,11 +14,11 @@ int main(void)
 	long long int llnum;
 	float f;
 
-	printf("Size of a char: %lu byte(s)\n", sizeof(ch));
-	printf("Size of an int: %lu byte(s)\n", sizeof(num));
-	printf("Size of a long: %lu byte(s)\n", sizeof(lnum));
-	printf("Size of a long long int: %lu byte(s)\n", sizeof(llnum));
-	printf("Size of a float: %lu byte(s)\n", sizeof(f));
+	printf("Size of a char: %d byte(s)\n", (int)sizeof(ch));
+	printf("Size of an int: %d byte(s)\n", (int)sizeof(num));
+	printf("Size of a long: %d byte(s)\n",(int)sizeof(lnum));
+	printf("Size of a long long int: %d byte(s)\n",(int)sizeof(llnum));
+	printf("Size of a float: %d byte(s)\n", (int)sizeof(f));
 
 	return (0);
 }
